refactor(core)!: move parseAssetId and createAssetId from util.assetId to coreToCml and cmlToCore
BREAKING CHANGE:
- parseAssetId was moved from util.assetId.ts to coreToCml.ts
- createAssetId was moved from util.assetId.ts to cmlToCore.ts
